requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.3
streamlit==1.29.0
requests==2.31.0python -m pip install --upgrade pip
